Refactor and move more logic for Sizers into SizerBase to share between GridSplitter and ContentSizer
- Share XAML template between all controls
  - Includes sharing VSM and logic to control that
- Centralized control manipulation logic to base class
  - Cleaned-up Keyboard handling code to be shared in base class using OnKeyDown override
  - Move OnManipulationDelta to SizerBase
- Share Automation Name and setup via XAML vs. code-behind
- Renamed ResizeDirection to Orientation in the base for ContentSizer and future controls, synced from GridSplitter with original ResizeDirection for polyfill (removed unneeded ContentSizerDirection/SizerResizeDirection enum as using system Orienation now)

TODO:
- Update/Fix AutomationPeer to generalize
- Provide OnLoaded virtual event?
- Figure out Content/ContentTemplate setup do we want this flexibility? Do we just have Content which is set to the TextBlock in Style as Default instead?
- Create Converter to set content for horizontal/vertical text in XAML (bound to Orientation)
- Update GripperCursor in Orientation changed callback
- Do we try inheriting from Thumb? Does that automatically solve the mouse drag back issue?
- Audit Sizer/Splitter/Gripper/Thumb naming???
