+++ Microsoft.Toolkit.Uwp.UI.Controls.Layout/Sizers/GridSplitter/GridSplitter.cs
@@ -18,10 +18,6 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         private GridResizeDirection _resizeDirection;
         private GridResizeBehavior _resizeBehavior;
 
-        private bool _pressed = false;
-        private bool _dragging = false;
-        private bool _pointerEntered = false;
-
         /// <summary>
         /// Gets the target parent grid from level
         /// </summary>
@@ -153,11 +149,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         /// Initializes a new instance of the <see cref="GridSplitter"/> class.
         /// </summary>
         public GridSplitter()
+            : base()
         {
-            DefaultStyleKey = typeof(GridSplitter);
             Loaded += GridSplitter_Loaded;
-            string automationName = "WCT_GridSplitter_AutomationName".GetLocalized("Microsoft.Toolkit.Uwp.UI.Controls.Layout/Resources");
-            AutomationProperties.SetName(this, automationName);
         }
 
         /// <inheritdoc />
@@ -167,66 +161,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
 
             // Unhook registered events
             Loaded -= GridSplitter_Loaded;
-            PointerEntered -= GridSplitter_PointerEntered;
-            PointerExited -= GridSplitter_PointerExited;
-            PointerPressed -= GridSplitter_PointerPressed;
-            PointerReleased -= GridSplitter_PointerReleased;
-            ManipulationStarted -= GridSplitter_ManipulationStarted;
-            ManipulationCompleted -= GridSplitter_ManipulationCompleted;
 
             // Register Events
             Loaded += GridSplitter_Loaded;
-            PointerEntered += GridSplitter_PointerEntered;
-            PointerExited += GridSplitter_PointerExited;
-            PointerPressed += GridSplitter_PointerPressed;
-            PointerReleased += GridSplitter_PointerReleased;
-            ManipulationStarted += GridSplitter_ManipulationStarted;
-            ManipulationCompleted += GridSplitter_ManipulationCompleted;
-        }
-
-        private void GridSplitter_PointerReleased(object sender, PointerRoutedEventArgs e)
-        {
-            _pressed = false;
-            VisualStateManager.GoToState(this, _pointerEntered ? "PointerOver" : "Normal", true);
-        }
-
-        private void GridSplitter_PointerPressed(object sender, PointerRoutedEventArgs e)
-        {
-            _pressed = true;
-            VisualStateManager.GoToState(this, "Pressed", true);
-        }
-
-        private void GridSplitter_PointerExited(object sender, PointerRoutedEventArgs e)
-        {
-            _pointerEntered = false;
-
-            if (!_pressed && !_dragging)
-            {
-                VisualStateManager.GoToState(this, "Normal", true);
-            }
-        }
-
-        private void GridSplitter_PointerEntered(object sender, PointerRoutedEventArgs e)
-        {
-            _pointerEntered = true;
-
-            if (!_pressed && !_dragging)
-            {
-                VisualStateManager.GoToState(this, "PointerOver", true);
-            }
-        }
-
-        private void GridSplitter_ManipulationCompleted(object sender, ManipulationCompletedRoutedEventArgs e)
-        {
-            _dragging = false;
-            _pressed = false;
-            VisualStateManager.GoToState(this, _pointerEntered ? "PointerOver" : "Normal", true);
-        }
-
-        private void GridSplitter_ManipulationStarted(object sender, ManipulationStartedRoutedEventArgs e)
-        {
-            _dragging = true;
-            VisualStateManager.GoToState(this, "Pressed", true);
         }
     }
 }
